pyproject.toml
[tool.poetry]
name = "servodoor-app"
version = "0.1.0"
description = "A simple GUI for controlling upto 18 (RC servo) t-maze doors."
license = "MIT"
authors = ["Will Dickson <wbd@caltech.edu>"]
readme = "README.md"
packages= [{include = "servodoor_app"}]
homepage = "https://github.com/willdickson/servodoor-app"
repository = "https://github.com/willdickson/servodoor-app"
classifiers = [ 
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: MIT License',
    'Development Status :: 3 - Alpha',
    ]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyside6 = "^6.5.2"
pyserial = "^3.5"
servodoor = {git="https://github.com/willdickson/servodoor"}
adafruit-ampy = "^1.1.0"

[tool.poetry.scripts]
servodoor = "servodoor_app.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
